chore: move deno unit tests to tests/deno
